+++ Source/Sitecore.FakeDb/Security/AccessControl/AllowAccessResult.cs
@@ -1,0 +1,12 @@
+﻿namespace Sitecore.FakeDb.Security.AccessControl
+{
+  using Sitecore.Security.AccessControl;
+
+  public class AllowAccessResult : AccessResult
+  {
+    public AllowAccessResult()
+      : base(AccessPermission.Allow, new AccessExplanation("Allow"))
+    {
+    }
+  }
+}